crea ordine

+++ src/main/java/control/CarrelloServlet.java
@@ -40,7 +40,7 @@ public class CarrelloServlet extends HttpServlet {
                 } catch (SQLException e) {
                     throw new RuntimeException(e);
                 }
-                totale+= (int) (prodotto.getPrezzo() * p.getQuantita());
+                totale+=prodotto.getPrezzo()*p.getQuantita();
                 prodottiStampa.add(prodotto);
             }
         }
